+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - registry: Normalize log messages
 - registry/storage/cache/redis: Add Prometheus metrics for Redis cache store
 - registry: Add support for a pprof monitoring server
+- registry: Add support for continuous profiling with Google Stackdriver
 
 ## [v2.9.1-gitlab] - 2020-05-05
 ## Added